goodix: use workqueue to load fb notifier
Blocking for a significant amount of time in the fb notifier can add
latency to panel power-on time when waking the device from sleep. To avoid
this, put all non-trivial fb notifier operations into a worker so that
they can execute asynchronously from the fb notifier.

Signed-off-by: Sultanxda <sultanxda@gmail.com>

rewrite code compatible for goodix new version

Signed-off-by: Gah0 <g0d3ph@gmail.com>

 Conflicts:
	drivers/input/fingerprint/goodix_fod/gf_spi.c
	drivers/input/fingerprint/goodix_fod/gf_spi.h